requirements.txt
# -*- coding: utf-8 -*-
"""
Created on Thu Jan  1 13:39:07 2026

@author: User
"""

aiogram==3.4.1
openai
Pillow
apscheduler
python-dotenv